Update index.html
If we include our script in the <head> normally, it will run as soon as it is downloaded, which will occur before the DOM has built the elements we want our script to run against. So our script wouldn´t be able to do anything. 
For that, we include it at the bottom of the <body>, but that would mean that the downloaded could potentially start later in the load process.